LPS-199873 search-experiences-service: Reorganize sort conversion for reusability
https://liferay.atlassian.net/browse/LPS-199873
